Fix server memory consumption
The signalfx drain accumulates data points in an array until the Flush
method is called. After that, the array should be reseted, but this was
not what was happening. Therefore, not only we the memory consumption
monotonically increase, but we also sent more and more amounts of data
to signalfx as the time passed.